% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{extract_chains_fme}
\alias{extract_chains_fme}
\title{Extract \code{FME} model chain objects into an R list}
\usage{
extract_chains_fme(julia_obj)
}
\arguments{
\item{julia_obj}{Julia environment object from \code{julia_setup} which has
also run the \code{FME} MCMC code.}
}
\value{
List of chain objects. Note the special unicode characters in the
Julia object are renamed using standard characters. Also note that parameters
associated with NULL data objects will be returned as empty arrays.
}
\description{
Custom structs like the returned chains object don't return properly in R.
This function extracts the elements into a list.
}
