% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{fit_model_fme}
\alias{fit_model_fme}
\title{Functional mixed effects model wrapper function}
\usage{
fit_model_fme(
  julia_obj,
  Yobs,
  Xfix = NULL,
  Xrand = NULL,
  Xcent = NULL,
  p = 20L,
  niter = 15000L,
  nburn = 5000L,
  nthin = 1L,
  save_random_effects = FALSE,
  save_theta = FALSE,
  a_sig = 3,
  b_sig = 1,
  a_tau = 3,
  b_tau = 1,
  a_lam = 100,
  b_lam = 30,
  v_fix = 1000,
  v_cent = 1000
)
}
\arguments{
\item{julia_obj}{A Julia environment from \code{julia_setup}.}

\item{Yobs, Xfix, Xrand, Xcent}{see documentation for \code{set_data_fme}}

\item{p, niter, nburn, nthin, save_random_effects}{see documentation for \code{set_config_fme}}

\item{a_sig, b_sig, a_tau, b_tau, a_lam, b_lam, v_fix, v_cent}{see documentation for \code{set_hyperparm_fme}}
}
\value{
MCMC chains object, see documentation for \code{mcmc_fme}.
}
\description{
Fits the functional mixed effects model. This allows the user to provide all
parameters at once rather than sequentially calling the interface functions.
}
