% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{set_data_fme}
\alias{set_data_fme}
\title{Set the \code{FME} data object in the Julia environment}
\usage{
set_data_fme(julia_obj, Yobs, Xfix = NULL, Xrand = NULL, Xcent = NULL)
}
\arguments{
\item{julia_obj}{Julia environment from \code{julia_setup}.}

\item{Yobs}{matrix containing curves/functions as columns.}

\item{Xfix}{matrix containing fixed effect covariates, with individuals contained
as columns and variables as rows (reverse of usual design matrices). Can be set
to \code{NULL} if only random effects are present.}

\item{Xrand}{matrix containing fixed effect covariates, with individuals contained
as columns and variables as rows (reverse of usual design matrices). Can be
set to \code{NULL} to perform functional regression without random effects.}

\item{Xcent}{matrix containing the mapping between random effects and their
hierarchical centered effect. Rows contain indicators for which random effects
are centered together. Can be set to \code{NULL} when random effects are also
missing.}
}
\value{
Invisibly returns list of data variables. The data
object \code{data} is saved in the \code{julia_obj} environment.
}
\description{
The argument names and values are save to the \code{julia_obj}
environment, then the data object is built. The values must
all return \code{TRUE} to \code{is.double}.
}
