% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup_julia}
\alias{setup_julia}
\title{Build Julia environment and load \code{FunctionalMixedEffects.jl} module.}
\usage{
setup_julia(install_julia = FALSE)
}
\arguments{
\item{install_julia}{Logical indicating if Julia should be installed in the
event that an install either is not installed or cannot be found on the path.}
}
\value{
A \code{JuliaCall} object which interfaces to a Julia session.
}
\description{
Build Julia environment and load \code{FunctionalMixedEffects.jl} module.
}
