% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{mcmc_fme}
\alias{mcmc_fme}
\title{Functional mixed effects MCMC}
\usage{
mcmc_fme(julia_obj)
}
\arguments{
\item{julia_obj}{Julia environment from \code{julia_setup}.}
}
\value{
Invisibly returns NULL. The resulting object \code{chains}
is saved in the \code{julia_obj} environment.
}
\description{
Fit the functional mixed model to data. \code{HyperParameter}, \code{Data},
and \code{OutputConfig} objects must be defined in the Julia environment..
}
