% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{set_config_fme}
\alias{set_config_fme}
\title{Set the \code{FME} config variable in the Julia environment.}
\usage{
set_config_fme(
  julia_obj,
  niter = 5000L,
  nburn = 1000L,
  nthin = 1L,
  save_random_effects = FALSE,
  save_theta = FALSE
)
}
\arguments{
\item{julia_obj}{Julia environment object from \code{julia_setup}.}

\item{niter, nburn, nthin}{MCMC iteration parameters. Strict integers are required.}

\item{save_random_effects}{Logical indicating whether the random effects
chain should be saved. Strict logical is required.}

\item{save_theta}{Logical indicating whether the individual coefficients
chain should be saved. Strict logical is required.}
}
\value{
Invisibly returns list of config variables. The config object \code{cfg}
is saved in the \code{julia_obj} environment.
}
\description{
The argument names and values are saved to the \code{julia_obj}
environment, then the config object is built.
}
