% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FME.R
\name{set_hyperparm_fme}
\alias{set_hyperparm_fme}
\title{Set the \code{FME} hyperparameters in the Julia environment.}
\usage{
set_hyperparm_fme(
  julia_obj,
  a_sig = 3,
  b_sig = 1,
  a_tau = 3,
  b_tau = 1,
  a_lam = 3,
  b_lam = 1,
  v_fix = 10000
)
}
\arguments{
\item{julia_obj}{Julia environment object from \code{julia_setup}.}

\item{a_sig, b_sig}{sigma priors. Must have \code{a > 2}}

\item{a_tau, b_tau}{tau priors. Must have \code{a > 2}}

\item{a_lam, b_lam}{lambda priors. Must have \code{a > 2}}

\item{v_fix}{fixed effect prior variance.}
}
\value{
Invisibly returns list of hyperparameter variables. The hyperparameter
object \code{hyps} is saved in the \code{julia_obj} environment.
}
\description{
The argument names and values are save to the \code{julia_obj}
environment, then the hyperparameter object is built. THe values must
all return \code{TRUE} to \code{is.double}.
}
